+++ integrated_channels/xapi/constants.py
@@ -1,0 +1,17 @@
+"""
+Constants for xAPI.
+"""
+from __future__ import absolute_import, unicode_literals
+
+# URI constants for xAPI verbs
+X_API_VERB_REGISTERED = 'http://adlnet.gov/expapi/verbs/registered'
+X_API_VERB_COMPLETED = 'http://adlnet.gov/expapi/verbs/completed'
+
+
+# URI constants for xAPI activities
+X_API_ACTIVITY_COURSE = 'http://adlnet.gov/expapi/activities/course'
+
+
+# Constants for edX grades
+MIN_SCORE = 0
+MAX_SCORE = 100